fix error with hibernate
