+++ src/main/resources/application.properties
@@ -4,6 +4,7 @@ spring.datasource.url=jdbc:mysql://localhost:3306/test?useSSL=false
 spring.datasource.username=root
 spring.datasource.password=PasswordTest
 spring.jpa.hibernate.ddl-auto=update
+spring.jpa.open-in-view=true
 spring.mvc.pathmatch.matching-strategy=ANT_PATH_MATCHER
 security.jwt.secret-key = thiskeyisjustfortestingpurposesdontuseitanywhereelsety
 spring.activemq.broker-url=tcp://localhost:61616
